chore:added server component
