+++ bciers/apps/reporting/src/app/bceidbusiness/industry_user_admin/reports/[version_id]/additional-reporting-data/page.tsx
@@ -1,26 +1,14 @@
 import { Suspense } from "react";
 import Loading from "@bciers/components/loading/SkeletonForm";
-import AdditionalReportingData from "@reporting/src/app/components/additionalInformation/AdditionalReportingData";
-import { getRegistrationPurpose } from "@reporting/src/app/utils/getRegistrationPurpose";
-
+import AdditionalReportingData from "@reporting/src/app/components/additionalInformation/additionalReportingData/AdditionalReportingData";
 interface PageProps {
   params: { version_id: number };
 }
 
 export default async function Page({ params }: PageProps) {
-  const versionId = params.version_id;
-  const registrationPurposes =
-    (await getRegistrationPurpose(versionId))?.registration_purposes || [];
-  const includeElectricityGenerated = registrationPurposes.includes(
-    "OBPS Regulated Operation",
-  );
-
   return (
     <Suspense fallback={<Loading />}>
-      <AdditionalReportingData
-        versionId={versionId}
-        includeElectricityGenerated={includeElectricityGenerated}
-      />
+      <AdditionalReportingData versionId={params.version_id} />
     </Suspense>
   );
 }
