+++ bciers/apps/reporting/src/data/jsonSchema/additionalReportingData/additionalReportingData.ts
@@ -144,12 +144,7 @@ export const additionalReportingDataUiSchema = {
   },
 };
 
-/**
- * Appends the electricity_generated field to the schema if includeElectricityGenerated is true.
- *
- * @returns The modified schema or the original schema if the flag is false.
- */
-export function getUpdatedSchema(): RJSFSchema {
+function buildSchemaWithElectricityGenerated(): RJSFSchema {
   return {
     ...additionalReportingDataSchema,
     properties: {
@@ -159,7 +154,7 @@ export function getUpdatedSchema(): RJSFSchema {
         title: "Additional data",
         properties: {
           electricity_generated: {
-            type: "string",
+            type: "number",
             title: "Electricity Generated",
           },
         },
@@ -167,3 +162,6 @@ export function getUpdatedSchema(): RJSFSchema {
     },
   };
 }
+
+export const additionalReportingDataWithElectricityGeneratedSchema: RJSFSchema =
+  buildSchemaWithElectricityGenerated();
